Phase B-6: kx-projection lib.rs split (Rule 3, behavior-preserving; D38 §2b)
Sixth Phase B refactor per Rule 3. **Rule 1 max-strict surface** (D38 §2b
9-cell recovery fold + STEP 5 ordering invariants). Pure move; zero
behavior change.

Before: src/lib.rs 1305 lines (3 public enums + RegisterMote + ProjectionError
+ private CommittedInfo/MoteInfo/DeclaredInfo/State + ~470-line impl State
+ Projection + ~430-line impl Projection + Snapshot + ~140-line impl
Snapshot + 3 private helper fns + 10 inline tests).

After: thin src/lib.rs 95 lines + 8 sibling modules:
- enums.rs (MoteState + AnomalyKind + PromotionState — public)
- register.rs (RegisterMote)
- errors.rs (ProjectionError)
- state.rs (CommittedInfo + MoteInfo + DeclaredInfo + State + impl State;
  internal machinery — all pub(crate) for cross-module access)
- projection.rs (Projection + impl Projection)
- snapshot.rs (Snapshot + impl Snapshot; pub(crate) state field for
  projection.rs's snapshot() constructor)
- helpers.rs (3 pub(crate) free fns: transitive_consumers_impl +
  ready_set_impl + promotion_state_impl)
- tests.rs (10 inline unit tests; #[cfg(test)] from lib.rs)

Visibility widenings (all pub(crate); no public API drift):
- State struct + fields + methods → pub(crate)
- CommittedInfo + MoteInfo + DeclaredInfo structs + fields → pub(crate)
- Snapshot::state field → pub(crate) (projection.rs constructs it)
- 3 helper free fns → pub(crate) (snapshot.rs + projection.rs call them)

Re-exports preserve every previously-public path (kx_projection::{
MoteState, AnomalyKind, PromotionState, RegisterMote, ProjectionError,
Projection, Snapshot}).

Local gates green on Apple Silicon (--all-features):
- cargo build -p kx-projection: clean
- cargo test --workspace --all-features: 645 + 8 (IDENTICAL)
- cargo test -p kx-projection: 10 + 19 + 14 + 13 + 2 = 58 (IDENTICAL)
- cargo clippy -p kx-projection --all-targets -- -D warnings: clean
- cargo fmt --all -- --check: clean

Next Phase B slice: kx-mote (1332 lines; FINAL B-7 — load-bearing trait
surface; the narrow waist of every downstream crate).
